requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
zope.interface==6.0

django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
django-hashid-field==3.3.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-recaptcha==3.0.0
django-extensions==3.2.1
easy-thumbnails==2.8.5
django-annoying==0.10.6
django-imagekit==4.1.0
django-user-sessions==2.0.0
django_user_agents==0.4.0
django-ipware==5.0.0
django-json-widget==1.1.1
fcm-django==1.0.12
django-rosetta==0.9.8
django-storages==1.13.2
django-autoslug==1.9.8
django-disable-cache-headers==0.0.1
django-anymail[mailgun]===9.1
django-debug-toolbar==3.8.1
django-graphql-jwt==0.3.4
django-oauth-toolkit==2.2.0
django-jsonform==2.16.1
drf-spectacular==0.26.1
django-ace==1.15.3
django-allauth==0.54.0